igor: fix colors with sort/filter
Colors between the table and reservations weren't lined up anymore.
Refactor things so that they match.
